Split managed runtime server helpers
